Bugfix: Skeleton Animation corruption when removing scene nodes. Removing lots of SceneNodes (>100) would trigger a defragmentation.
The defragmentation performs the following:
1. Move all memory around until everything is contiguous again. For example take N nodes occupying S slots from [0; S) and now make them occupy [0; N) where N < S
2. Initialize with default values the now empty space in range [N; S)
3. Update all the pointer references (since addresses to transforms have changed)

slotsRecreated was performing step 2; while performCleanup & applyRebase perform step 3.
However for SceneNodes slotsRecreated also was performing part of step 3.

This caused that when enough SceneNodes were removed, the defragmentation would end up having the skeletons disappear or glitch.
What happened was:
1. slotsRecreated gets called, performing step 2
2. SceneNode::_callMemoryChangeListeners would be called (this is step 3!!!)
3. Bone::_setNodeParent would be called. However it would read the old transform address; as the address will be later updated in performCleanup.

A quick solution would be to call slotsRecreated after performCleanup. However to make the code much more clear, slotsRecreated was renamed to initializeEmptySlots; and we now ensure it only performs step 3. The bit of step 2 (_callMemoryChangeListeners) was moved to performCleanup & applyRebase.

Thanks to user zxz for discovering this bug and providing a fix.

--HG--
branch : v2-1